requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
python-multipart==0.0.6

# AI & ML
langchain==0.2.0
langchain-openai==0.1.8
langchain-community==0.2.0
langgraph==0.0.19

# Vector Database
pymilvus==2.3.6
langchain-milvus==0.0.1
chromadb==0.4.22
langchain-chroma==0.1.0

# Data Processing
python-docx==1.1.0
pdfplumber==0.10.3
python-magic==0.4.27
unstructured[local-inference]==0.12.5

# Database
sqlalchemy==2.0.25
alembic==1.13.1
aiosqlite==0.19.0

# Utilities
python-dotenv==1.0.0
loguru==0.7.2
tenacity==8.2.3
httpx==0.26.0
aiofiles==23.2.1

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0

# Development
black==23.12.1
isort==5.13.2
flake8==7.0.0
mypy==1.8.0
